backend/requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cron-descriptor==1.2.35
Django==4.1.7
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-timezone-field==5.0
django-userforeignkey==0.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
kombu==5.2.4
pika==1.3.1
Pillow==9.4.0
prompt-toolkit==3.0.38
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
redis==4.5.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.6
